PMT calculations for cash flow calculations

+++ src/financialAnalysis/IncomeStatement.java
@@ -1,6 +1,6 @@
 package financialAnalysis;
 
 public class IncomeStatement {
-    private float grossRentalRevenue;
-    private float expenses;
+    private double grossRentalRevenue;
+    private double expenses;
 }
